+++ src/Cache/CacheKeyAndTagsGenerator.php
@@ -49,10 +49,8 @@ class CacheKeyAndTagsGenerator implements CacheKeyAndTags
     /**
      * @param  int|string|null $id
      */
-    public function parentTag($id, ResolveInfo $resolveInfo, ?string $parentName = null): string
+    public function parentTag(string $parentName, $id): string
     {
-        $parentName = $parentName ?? $resolveInfo->parentType->name;
-
         return implode(self::SEPARATOR, [
             self::PREFIX,
             $parentName,
@@ -63,11 +61,8 @@ class CacheKeyAndTagsGenerator implements CacheKeyAndTags
     /**
      * @param  int|string|null $id
      */
-    public function fieldTag($id, ResolveInfo $resolveInfo, ?string $parentName = null, ?string $fieldName = null): string
+    public function fieldTag(string $parentName, $id, string $fieldName): string
     {
-        $parentName = $parentName ?? $resolveInfo->parentType->name;
-        $fieldName = $fieldName ?? $resolveInfo->fieldName;
-
         return implode(self::SEPARATOR, [
             self::PREFIX,
             $parentName,