+++ src/Cache/ClearCacheDirective.php
@@ -89,8 +89,8 @@ GRAPHQL;
 
                 foreach ((array) $idOrIds as $id) {
                     $tag = is_string($field)
-                        ? $this->cacheKeyAndTags->fieldTag($id, $resolveInfo, $type, $field)
-                        : $this->cacheKeyAndTags->parentTag($id, $resolveInfo, $type);
+                        ? $this->cacheKeyAndTags->fieldTag($type, $id, $field)
+                        : $this->cacheKeyAndTags->parentTag($type, $id);
 
                     $this->cacheRepository->tags([$tag])->flush();
                 }